requirements.txt
django==1.11.3
virtualenv==15.1.0
pymysql==0.7.9
scrapy==1.4.0
scikit-learn==0.18.2
beautifulsoup4==4.6.0
mecab-python3==0.7
numpy==1.13.1
scipy==0.19.1
nltk==3.2.4
dill==0.2.7
